refactor: rename app layout folder
