+++ cats-pinterest/src/app/layouts/AppLayout.tsx
