+++ cats-pinterest/src/app/routes/router.tsx
@@ -1,7 +1,7 @@
 import { HashRouter, Navigate, Route, Routes } from "react-router";
 import { MainPage } from "../../pages/MainPage/MainPage";
 import { FavoritesPage } from "../../pages/FavoritesPage/FavoritesPage";
-import { AppLayout } from "../layouts/AppLayouts";
+import { AppLayout } from "../layouts/AppLayout";
 import { ROUTES } from "../../shared/constants/routes";
 
 export function AppRouter() {
